settings for analytics
