+++ cmd/arkgopool/goreleaser.yml
@@ -1,0 +1,10 @@
+# .goreleaser.yml
+# Build customization
+builds:
+  - binary: arkgopool
+    goos:
+      - windows
+      - darwin
+      - linux
+    goarch:
+      - amd64
