+++ cmd/arkgopool/payouts.go
@@ -7,10 +7,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/dghubble/sling"
 	"github.com/fatih/color"
 	"github.com/kristjank/ark-go/arkcoin"
-	"github.com/kristjank/ark-go/cmd/model"
 	"github.com/kristjank/ark-go/core"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -349,7 +347,9 @@ func SendPayments(silent bool) {
 
 		splitAndDeliverPayload(payload)
 		commitTx(dbtx)
-		go sendStatisticsData(&payrec)
+		if viper.GetBool("client.statistics") {
+			go sendStatisticsData(&payrec)
+		}
 
 		fmt.Println("Automated Payment complete. Please check the logs folder... ")
 		log.Info("Automated Payment complete. Please check the logs folder... ")
@@ -364,72 +364,6 @@ func SendPayments(silent bool) {
 	}
 }
 
-func splitAndDeliverPayload(payload core.TransactionPayload) {
-	//calculating number of chunks (based on 20tx in one chunk to send to one peer)
-	payoutsFolderName := createLogFolder()
-	var divided [][]*core.Transaction
-	numPeers := len(payload.Transactions) / 20
-	if numPeers == 0 {
-		numPeers = 1
-	}
-	chunkSize := (len(payload.Transactions) + numPeers - 1) / numPeers
-	if chunkSize == 0 {
-		chunkSize = 1
-	}
-
-	//sliptting the payload to number of needed peers
-	for i := 0; i < len(payload.Transactions); i += chunkSize {
-		end := i + chunkSize
-		if end > len(payload.Transactions) {
-			end = len(payload.Transactions)
-		}
-		divided = append(divided, payload.Transactions[i:end])
-	}
-	//end of spliting transactions
-
-	var tmpPayload core.TransactionPayload
-	splitcout := 0
-	for chunkIx, h := range divided {
-		tmpPayload.Transactions = h
-		splitcout += len(h)
-
-		deliverPayloadThreaded(tmpPayload, chunkIx, payoutsFolderName)
-
-	}
-	if splitcout != len(payload.Transactions) {
-		log.Info("TX spliting not OK")
-	}
-}
-
-func deliverPayloadThreaded(tmpPayload core.TransactionPayload, chunkIx int, logFolder string) {
-	numberOfPeers2MultiBroadCastTo := viper.GetInt("client.multibroadcast")
-	if numberOfPeers2MultiBroadCastTo > 15 {
-		numberOfPeers2MultiBroadCastTo = 15
-		log.Warn("Max broadcast number too high - set by user, reseting to value 15")
-	}
-	log.Info("Starting multibroadcast/multithreaded parallel payout to ", numberOfPeers2MultiBroadCastTo, " number of peers")
-	peers := arkclient.GetRandomXPeers(numberOfPeers2MultiBroadCastTo)
-	for i := 0; i < numberOfPeers2MultiBroadCastTo; i++ {
-		wg.Add(1)
-
-		//treaded function
-		go func(tmpPayload core.TransactionPayload, peer core.Peer, chunkIx int, logFolder string) {
-			defer wg.Done()
-			filename := fmt.Sprintf("log/%s/Batch_%02d_Peer%s.csv", logFolder, chunkIx, peer.IP)
-
-			arkTmpClient := core.NewArkClientFromPeer(peer)
-			res, _, _ := arkTmpClient.PostTransaction(tmpPayload)
-			if res.Success {
-				color.Set(color.FgHiGreen)
-				log2csv(tmpPayload, res.TransactionIDs, filename, "OK")
-			} else {
-				color.Set(color.FgHiRed)
-				log2csv(tmpPayload, nil, filename, res.Error)
-			}
-		}(tmpPayload, peers[i], chunkIx, logFolder)
-	}
-}
-
 func calcFidelity(element core.DelegateDataProfit) float64 {
 	fAmount2Send := element.EarnedAmountXX
 	//FIDELITY
@@ -441,25 +375,6 @@ func calcFidelity(element core.DelegateDataProfit) float64 {
 	}
 
 	return fAmount2Send
-}
-
-type postStatsResponse struct {
-	Success bool   `json:"success,omitempty"`
-	LogID   int    `json:"logID,omitempty"`
-	Error   string `json:"error,omitempty"`
-}
-
-func sendStatisticsData(payRec *model.PaymentRecord) {
-	response := new(postStatsResponse)
-	error := new(postStatsResponse)
-
-	statsBase := sling.New().Base("http://164.8.251.91:54010").Client(nil).Add("Content-Type", "application/json")
-	//statsBase := sling.New().Base("http://127.0.0.1:54010").Client(nil).Add("Content-Type", "application/json")
-	resp, err := statsBase.New().Post("log/payment").BodyJSON(payRec).Receive(response, error)
-
-	if err != nil {
-		log.Error("Error sending statistics data", resp, err)
-	}
 }
 
 //SendBonusPayment based on parameters in config.toml
@@ -560,7 +475,10 @@ func SendBonusPayment(iAmount int, txDesc string) {
 
 		splitAndDeliverPayload(payload)
 		commitTx(dbtx)
-		go sendStatisticsData(&payrec)
+
+		if viper.GetBool("client.statistics") {
+			go sendStatisticsData(&payrec)
+		}
 
 		fmt.Println("Automated Payment complete. Please check the logs folder... ")
 		log.Info("Automated Payment complete. Please check the logs folder... ")
