+++ cmd/arkgostats/goreleaser.yml
@@ -1,0 +1,10 @@
+# .goreleaser.yml
+# Build customization
+builds:
+  - binary: arkgostats
+    goos:
+      - windows
+      - darwin
+      - linux
+    goarch:
+      - amd64
